rc.boot: Remove dmesg restrict and kill eudev.
Restricting dmesg can be done permanently after boot by
the user and removing the code from the init script will
speed up boot.

We now kill eudev once the boot is complete as it the
daemon is optional on a KISS machine. If a user would like
to run the daemon afterwards a proper runit service can
simply be started.